feat: added utils for parse amenities and others
