+++ utils/parseAmenities.js
@@ -1,0 +1,11 @@
+import parseIcons from "~/utils/parseIcons"
+
+function parseAmenities(obj) {
+    return obj.amenities.map((amenity, index )=> {
+        return {
+            title: amenity,
+            icon: parseIcons(obj)[index]
+        }
+    })
+}
+export default parseAmenities
